Break up init
